requirements.txt
fastapi~=0.112.1
uvicorn~=0.30.6
typer~=0.12.4
strawberry-graphql~=0.237.3
SQLAlchemy~=2.0.32
python-dotenv~=1.0.1
alembic~=1.13.2
asyncpg~=0.29.0
cryptography~=43.0.0
aiohttp~=3.10.5
pydantic~=2.8.2
pydantic-settings~=2.4.0
starlette~=0.38.2
redis~=5.0.8